Moved ShipkitConfiguration to 'configuration' package
Needed for consistency with where currently the configuration plugin is stored.
