+++ src/main/groovy/org/shipkit/gradle/configuration/ShipkitConfiguration.java
@@ -1,4 +1,4 @@
-package org.shipkit.gradle;
+package org.shipkit.gradle.configuration;
 
 import org.gradle.api.GradleException;
 import org.gradle.api.logging.Logger;
