+++ src/main/groovy/org/shipkit/internal/gradle/configuration/ShipkitConfigurationPlugin.java
@@ -4,7 +4,7 @@ import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.plugins.ObjectConfigurationAction;
-import org.shipkit.gradle.ShipkitConfiguration;
+import org.shipkit.gradle.configuration.ShipkitConfiguration;
 import org.shipkit.internal.gradle.init.InitPlugin;
 import org.shipkit.internal.gradle.version.VersioningPlugin;
 import org.shipkit.internal.version.VersionInfo;
