+++ src/main/groovy/org/shipkit/internal/gradle/contributors/ContributorsPlugin.java
@@ -3,7 +3,7 @@ package org.shipkit.internal.gradle.contributors;
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.shipkit.gradle.ShipkitConfiguration;
+import org.shipkit.gradle.configuration.ShipkitConfiguration;
 import org.shipkit.gradle.notes.FetchContributorsTask;
 import org.shipkit.internal.gradle.configuration.ShipkitConfigurationPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
