+++ src/main/groovy/org/shipkit/internal/gradle/git/GitPush.java
@@ -2,7 +2,7 @@ package org.shipkit.internal.gradle.git;
 
 import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
-import org.shipkit.gradle.ShipkitConfiguration;
+import org.shipkit.gradle.configuration.ShipkitConfiguration;
 import org.shipkit.gradle.git.GitPushTask;
 import org.shipkit.internal.exec.DefaultProcessRunner;
