+++ src/main/groovy/org/shipkit/internal/gradle/notes/ReleaseNotesPlugin.java
@@ -3,11 +3,11 @@ package org.shipkit.internal.gradle.notes;
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.shipkit.gradle.ShipkitConfiguration;
+import org.shipkit.gradle.configuration.ShipkitConfiguration;
+import org.shipkit.gradle.notes.FetchContributorsTask;
 import org.shipkit.gradle.notes.FetchReleaseNotesTask;
 import org.shipkit.gradle.notes.UpdateReleaseNotesTask;
 import org.shipkit.internal.gradle.configuration.ShipkitConfigurationPlugin;
-import org.shipkit.gradle.notes.FetchContributorsTask;
 import org.shipkit.internal.gradle.contributors.ContributorsPlugin;
 import org.shipkit.internal.gradle.git.GitPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
