+++ src/main/groovy/org/shipkit/internal/gradle/release/ReleaseNeededPlugin.java
@@ -3,9 +3,9 @@ package org.shipkit.internal.gradle.release;
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.shipkit.gradle.ShipkitConfiguration;
-import org.shipkit.gradle.release.ReleaseNeededTask;
+import org.shipkit.gradle.configuration.ShipkitConfiguration;
 import org.shipkit.gradle.java.ComparePublicationsTask;
+import org.shipkit.gradle.release.ReleaseNeededTask;
 import org.shipkit.internal.gradle.configuration.ShipkitConfigurationPlugin;
 import org.shipkit.internal.gradle.git.GitBranchPlugin;
 import org.shipkit.internal.gradle.java.ComparePublicationsPlugin;
